backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib
bcrypt==3.2.0
python-multipart
websockets